refactor(config): rename layout dimensions module
Use an explicit module name for layout dimension constants so the config file cannot be confused with Next.js layout files.

Refs: #1724
Made-with: Cursor
