+++ src/config/index.ts
@@ -3,7 +3,7 @@ export * from './database';
 export * from './externalLinks';
 export * from './feed';
 export * from './forms';
-export * from './layout';
+export * from './layoutDimensions';
 export * from './network';
 export * from './nexus';
 export * from './posts';
